Updated the package name so it's not camel case
