+++ ComposeAdvanced/app/src/main/java/com/example/android/wearable/composeadvanced/presentation/ui/userinput/SliderScreen.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.android.wearable.composeadvanced.presentation.ui.userInputComponents
+package com.example.android.wearable.composeadvanced.presentation.ui.userinput
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -22,15 +22,14 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.wear.compose.material.Stepper
-import androidx.wear.compose.material.Text
+import androidx.wear.compose.material.InlineSlider
 
 /**
- * Displays a Stepper, which allows users to make a selection from a range of values.
- * https://developer.android.com/reference/kotlin/androidx/wear/compose/material/package-summary#Stepper(kotlin.Int,kotlin.Function1,androidx.compose.ui.Modifier,kotlin.ranges.IntProgression,kotlin.Function0,kotlin.Function0,androidx.compose.ui.graphics.Color,androidx.compose.ui.graphics.Color,androidx.compose.ui.graphics.Color,kotlin.Function1)
+ * Displays a Slider, which allows users to make a selection from a range of values.
+ * https://developer.android.com/reference/kotlin/androidx/wear/compose/material/package-summary#InlineSlider(kotlin.Float,kotlin.Function1,kotlin.Int,androidx.compose.ui.Modifier,kotlin.Boolean,kotlin.ranges.ClosedFloatingPointRange,kotlin.Boolean,kotlin.Function0,kotlin.Function0,androidx.wear.compose.material.InlineSliderColors)
  */
 @Composable
-fun StepperScreen(
+fun SliderScreen(
     displayValue: Int,
     onValueChange: (Int) -> Unit
 ) {
@@ -40,10 +39,11 @@ fun StepperScreen(
             .padding(top = 10.dp),
         verticalArrangement = Arrangement.Center
     ) {
-        Stepper(
+        InlineSlider(
             value = displayValue,
             onValueChange = onValueChange,
-            valueProgression = 1..10
-        ) { Text("Value: $displayValue") }
+            valueProgression = 1..10,
+            segmented = true
+        )
     }
 }
