+++ ComposeAdvanced/app/src/main/java/com/example/android/wearable/composeadvanced/presentation/ui/userinput/StepperScreen.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.android.wearable.composeadvanced.presentation.ui.userInputComponents
+package com.example.android.wearable.composeadvanced.presentation.ui.userinput
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -22,14 +22,15 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.wear.compose.material.InlineSlider
+import androidx.wear.compose.material.Stepper
+import androidx.wear.compose.material.Text
 
 /**
- * Displays a Slider, which allows users to make a selection from a range of values.
- * https://developer.android.com/reference/kotlin/androidx/wear/compose/material/package-summary#InlineSlider(kotlin.Float,kotlin.Function1,kotlin.Int,androidx.compose.ui.Modifier,kotlin.Boolean,kotlin.ranges.ClosedFloatingPointRange,kotlin.Boolean,kotlin.Function0,kotlin.Function0,androidx.wear.compose.material.InlineSliderColors)
+ * Displays a Stepper, which allows users to make a selection from a range of values.
+ * https://developer.android.com/reference/kotlin/androidx/wear/compose/material/package-summary#Stepper(kotlin.Int,kotlin.Function1,androidx.compose.ui.Modifier,kotlin.ranges.IntProgression,kotlin.Function0,kotlin.Function0,androidx.compose.ui.graphics.Color,androidx.compose.ui.graphics.Color,androidx.compose.ui.graphics.Color,kotlin.Function1)
  */
 @Composable
-fun SliderScreen(
+fun StepperScreen(
     displayValue: Int,
     onValueChange: (Int) -> Unit
 ) {
@@ -39,11 +40,10 @@ fun SliderScreen(
             .padding(top = 10.dp),
         verticalArrangement = Arrangement.Center
     ) {
-        InlineSlider(
+        Stepper(
             value = displayValue,
             onValueChange = onValueChange,
-            valueProgression = 1..10,
-            segmented = true
-        )
+            valueProgression = 1..10
+        ) { Text("Value: $displayValue") }
     }
 }
